Add modal for search on mobile version
