+++ web/components/Search/SearchInput/SearchInput.scss
@@ -1,0 +1,24 @@
+.search-form {
+  gap: 1rem;
+  display: flex;
+  width: 100%;
+  max-width: 600px;
+}
+.search-form .search-input {
+  position: relative;
+  flex: 1 0 auto;
+  height: 100%;
+}
+.search-form > button {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  $size: 50px;
+  flex: 0 0 70px;
+  height: 100%;
+  svg {
+    $size: 35px;
+    height: $size;
+    width: $size;
+  }
+}
